refactor II
